Gracefully handle an empty table list in sql->drop()

+++ lib/Drush/Sql/SqlBase.php
@@ -227,8 +227,10 @@ class SqlBase {
   }
 
   public function drop($tables) {
-    $sql = 'DROP TABLE '. implode(', ', $tables);
-    return $this->query($sql);
+    if ($tables) {
+      $sql = 'DROP TABLE '. implode(', ', $tables);
+      return $this->query($sql);
+    }
   }
 
   /**
